requirements.txt
pyspark==2.4.5
py4j==0.10.7
pyspark-stubs==2.4.0.post8
pyyaml==5.3.1
pyspark==2.4.5
dataclasses==0.6
pymongo==3.10.1
# PYTHONUNBUFFERED=1;PYSPARK_PYTHON=/usr/local/bin/python3.7;PYSPARK_DRIVER_PYTHON=/usr/local/bin/python3.7
